+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/pokemon/gen8/CorvisquireModel.kt
@@ -1,0 +1,147 @@
+package com.cobblemon.mod.common.client.render.models.blockbench.pokemon.gen8
+
+import com.cobblemon.mod.common.client.render.models.blockbench.frame.BiWingedFrame
+import com.cobblemon.mod.common.client.render.models.blockbench.frame.BipedFrame
+import com.cobblemon.mod.common.client.render.models.blockbench.frame.HeadedFrame
+import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPose
+import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPoseableModel
+import com.cobblemon.mod.common.client.render.models.blockbench.pose.TransformedModelPart
+import com.cobblemon.mod.common.client.render.models.blockbench.wavefunction.parabolaFunction
+import com.cobblemon.mod.common.client.render.models.blockbench.wavefunction.sineFunction
+import com.cobblemon.mod.common.entity.PoseType
+import com.cobblemon.mod.common.util.math.geometry.toRadians
+import net.minecraft.client.model.ModelPart
+import net.minecraft.util.math.Vec3d
+
+class CorvisquireModel (root: ModelPart) : PokemonPoseableModel(), HeadedFrame, BipedFrame, BiWingedFrame {
+    override val rootPart = root.registerChildWithAllChildren("corvisquire")
+    override val leftWing = getPart("wing_left")
+    override val rightWing = getPart("wing_right")
+    override val leftLeg = getPart("leg_left")
+    override val rightLeg = getPart("leg_right")
+    override val head = getPart("head")
+
+
+    override val portraitScale = 3.5F
+    override val portraitTranslation = Vec3d(-0.1, -2.1, 0.0)
+
+    override val profileScale = 1.2F
+    override val profileTranslation = Vec3d(0.0, -0.01, 0.0)
+
+
+    lateinit var stand: PokemonPose
+    lateinit var walk: PokemonPose
+    lateinit var hover: PokemonPose
+    lateinit var fly: PokemonPose
+
+    override fun registerPoses() {
+
+        val blink = quirk("blink") { bedrockStateful("corvisquire", "blink").setPreventsIdle(false) }
+        stand = registerPose(
+            poseName = "standing",
+            poseTypes = PoseType.SHOULDER_POSES + PoseType.UI_POSES + PoseType.STAND,
+            transformTicks = 10,
+            quirks = arrayOf(blink),
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("corvisquire", "ground_idle")
+            )
+        )
+
+        hover = registerPose(
+            poseName = "hover",
+            poseType = PoseType.HOVER,
+            transformTicks = 10,
+            quirks = arrayOf(blink),
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("corvisquire", "air_idle")
+            )
+        )
+
+        fly = registerPose(
+            poseName = "fly",
+            poseType = PoseType.FLY,
+            transformTicks = 10,
+            quirks = arrayOf(blink),
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("corvisquire", "air_fly")
+            )
+        )
+
+        walk = registerPose(
+            poseName = "walking",
+            poseType = PoseType.WALK,
+            transformTicks = 10,
+            quirks = arrayOf(blink),
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("corvisquire", "ground_idle"),
+                rootPart.translation(
+                    function = parabolaFunction(
+                        peak = -4F,
+                        period = 0.4F
+                    ),
+                    timeVariable = { state, _, _ -> state?.animationSeconds },
+                    axis = TransformedModelPart.Y_AXIS
+                ),
+                head.translation(
+                    function = sineFunction(
+                        amplitude = (-20F).toRadians(),
+                        period = 1F,
+                        verticalShift = (-10F).toRadians()
+                    ),
+                    axis = TransformedModelPart.X_AXIS,
+                    timeVariable = { state, _, _ -> state?.animationSeconds }
+                ),
+                leftLeg.rotation(
+                    function = parabolaFunction(
+                        tightness = -20F,
+                        phaseShift = 0F,
+                        verticalShift = (30F).toRadians()
+                    ),
+                    axis = TransformedModelPart.X_AXIS,
+                    timeVariable = { _, _, ageInTicks -> ageInTicks / 20 },
+                ),
+                rightLeg.rotation(
+                    function = parabolaFunction(
+                        tightness = -20F,
+                        phaseShift = 0F,
+                        verticalShift = (30F).toRadians()
+                    ),
+                    axis = TransformedModelPart.X_AXIS,
+                    timeVariable = { _, _, ageInTicks -> ageInTicks / 20 },
+                ),
+                wingFlap(
+                    flapFunction = sineFunction(
+                        amplitude = (-5F).toRadians(),
+                        period = 0.4F,
+                        phaseShift = 0.00F,
+                        verticalShift = (-20F).toRadians()
+                    ),
+                    timeVariable = { state, _, _ -> state?.animationSeconds },
+                    axis = TransformedModelPart.Z_AXIS
+                ),
+                rightWing.translation(
+                    function = parabolaFunction(
+                        tightness = -10F,
+                        phaseShift = 30F,
+                        verticalShift = (25F).toRadians()
+                    ),
+                    axis = TransformedModelPart.Y_AXIS,
+                    timeVariable = { _, _, ageInTicks -> ageInTicks / 20 },
+                ),
+                leftWing.translation(
+                    function = parabolaFunction(
+                        tightness = -10F,
+                        phaseShift = 30F,
+                        verticalShift = (25F).toRadians()
+                    ),
+                    axis = TransformedModelPart.Y_AXIS,
+                    timeVariable = { _, _, ageInTicks -> ageInTicks / 20 },
+                ),
+            )
+        )
+    }
+}