+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/pokemon/gen8/SirfetchdModel.kt
@@ -1,0 +1,58 @@
+package com.cobblemon.mod.common.client.render.models.blockbench.pokemon.gen8
+
+import com.cobblemon.mod.common.client.render.models.blockbench.animation.BipedWalkAnimation
+import com.cobblemon.mod.common.client.render.models.blockbench.frame.BipedFrame
+import com.cobblemon.mod.common.client.render.models.blockbench.frame.HeadedFrame
+import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPose
+import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPoseableModel
+import com.cobblemon.mod.common.entity.PoseType
+import net.minecraft.client.model.ModelPart
+import net.minecraft.util.math.Vec3d
+
+class SirfetchdModel (root: ModelPart) : PokemonPoseableModel(), HeadedFrame, BipedFrame {
+    override val rootPart = root.registerChildWithAllChildren("sirfetchd")
+    override val head = getPart("head")
+
+    override val leftLeg = getPart("leg_left")
+    override val rightLeg = getPart("leg_right")
+
+    override val portraitScale = 2.2F
+    override val portraitTranslation = Vec3d(-0.35, -1.0, 0.0)
+
+    override val profileScale = 1.1F
+    override val profileTranslation = Vec3d(-0.1, 0.1, 0.0)
+
+    lateinit var standing: PokemonPose
+    lateinit var walk: PokemonPose
+
+    override fun registerPoses() {
+        val blink = quirk("blink") { bedrockStateful("sirfetchd", "blink").setPreventsIdle(false) }
+        standing = registerPose(
+            poseName = "standing",
+            poseTypes = PoseType.STATIONARY_POSES + PoseType.UI_POSES,
+            transformTicks = 10,
+            quirks = arrayOf(blink),
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("sirfetchd", "ground_idle")
+            )
+        )
+
+        walk = registerPose(
+            poseName = "walk",
+            poseTypes = PoseType.MOVING_POSES,
+            transformTicks = 5,
+            quirks = arrayOf(blink),
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("sirfetchd", "ground_idle"),
+                BipedWalkAnimation(this, periodMultiplier = 0.6F, amplitudeMultiplier = 0.9F),
+            )
+        )
+    }
+
+//    override fun getFaintAnimation(
+//        pokemonEntity: PokemonEntity,
+//        state: PoseableEntityState<PokemonEntity>
+//    ) = if (state.isPosedIn(standing, walk)) bedrockStateful("sirfetchd", "faint") else null
+}